Backend/pharmacy-api/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv
fhirclient>=3.2.0
fhir.resources>=6.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
fastapi-users[sqlalchemy]>=12.0.0
email-validator>=2.0.0
fastapi-users-db-sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic-settings>=2.0.0
